feat(Footer.vue, Background.vue): add new components and assets (relaunch-2024)
This commit includes changes to `Footer.vue` and a new component `Background.vue`.

In `Footer.vue` the structure was updated and now includes the `Background.vue` component for enhanced styling.

The new `Background.vue` component receives a number of props to handle the positioning, height, width, and source of a background image.

Additionally, two images `footer-top.svg` and `box.png` were added to the assets.

These changes contribute to the application's relaunch for 2024.