feat: add Menu component to app.vue
A new Menu component has been added to the app.vue file. This includes the implementation of a navigation bar with responsive design and different links to different sections of the website. The menu button has been added with toggle functionality to open and close the menu. This commit also includes the addition of a new file, components/Menu.vue, where the template and script for the new component reside.