feat: add @nuxtjs/web-vitals for web performance tracking
This commit includes adding the @nuxtjs/web-vitals package to the jott-media project. This package is used for measuring, collecting, and reporting web vitals to improve the performance of the website. The package is added to the package.json and package-lock.json files. Additionally, the config file nuxt.config.ts has been updated to include the @nuxtjs/web-vitals module and its configurations.