feat(blog): update component names and structure in blog page
This commit includes the renaming of the 'Uarticles' components to 'UPage' and 'UarticlesBody' to 'UPageBody'. It also modifies the author links to display the author's name without changing it to lower case. The commit further adds more fields to the query collection and sets up a watcher on the locale.